Refactored portal offer helper
no ref
